Move asset builder functions
